Reorganized persistence package
