+++ src/main/java/com/gznznzjsn/outstagram/persistence/repository/AccountRepository.java
@@ -1,4 +1,4 @@
-package com.gznznzjsn.outstagram.repository;
+package com.gznznzjsn.outstagram.persistence.repository;
 
 import com.gznznzjsn.outstagram.model.node.Account;
 import org.apache.ibatis.annotations.Mapper;
