+++ src/main/resources/db/mybatis/mapper/subscription-repository.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8" ?>
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd" >
 
-<mapper namespace="com.gznznzjsn.outstagram.repository.SubscriptionRepository">
+<mapper namespace="com.gznznzjsn.outstagram.persistence.repository.SubscriptionRepository">
 
     <insert id="create">
         match (a:ACCOUNT), (b:ACCOUNT)
